+++ src/features/comment/useCommentActions.ts
@@ -28,7 +28,7 @@ import {
   getModIcon,
 } from "#/features/moderation/useCanModerate";
 import useCommentModActions from "#/features/moderation/useCommentModActions";
-import { useSharePostComment } from "#/features/post/actions/useSharePostComment";
+import { useSharePostComment } from "#/features/shared/useSharePostComment";
 import { getShareIcon } from "#/helpers/device";
 import {
   getHandle,
