+++ src/features/post/shared/usePostActions.tsx
@@ -28,7 +28,7 @@ import {
   getModIcon,
 } from "#/features/moderation/useCanModerate";
 import usePostModActions from "#/features/moderation/usePostModActions";
-import { useSharePostComment } from "#/features/post/actions/useSharePostComment";
+import { useSharePostComment } from "#/features/shared/useSharePostComment";
 import { resolveObject } from "#/features/resolve/resolveSlice";
 import { getShareIcon } from "#/helpers/device";
 import { getCrosspostUrl, getHandle, getRemoteHandle } from "#/helpers/lemmy";
