+++ src/features/shared/sliding/internal/impl/VotableActionsImpl.tsx
@@ -14,7 +14,7 @@ import { scrollCommentIntoViewIfNeeded } from "#/features/comment/inTree/Comment
 import useCollapseRootComment from "#/features/comment/inTree/useCollapseRootComment";
 import { markRead } from "#/features/inbox/inboxSlice";
 import { getCanModerate } from "#/features/moderation/useCanModerate";
-import { useSharePostComment } from "#/features/post/actions/useSharePostComment";
+import { useSharePostComment } from "#/features/shared/useSharePostComment";
 import { savePost, voteOnPost } from "#/features/post/postSlice";
 import {
   isInboxItem,
